+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Spell tooltip channeling time 
 - Rework summoned entity sound data
+- Added invulnerability recheck after `PLAYER_DAMAGE_INCOMING.PRE` event
 
 # 1.9.11